Commit 2.5

+++ demo/src/main/java/com/example/demo/services/CiudadService.java
@@ -1,8 +1,10 @@
 package com.example.demo.services;
 
-import com.example.demo.entities.Ciudad;
+import com.example.demo.dtos.IdsDtos;
+import com.example.demo.entities.Mysql.Ciudad;
 import com.example.demo.repository.Mysql.CiudadRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,15 +13,22 @@ import java.util.Optional;
 @Service
 @Transactional
 public class CiudadService {
+    @Autowired
+    private LogService logService;
 
     @Autowired
     private CiudadRepository ciudadRepository;
 
-    public void deleteById(Integer id) {
-        ciudadRepository.deleteById(id);
-    }
-
-    public Optional<Ciudad> findById(Integer id) {
-        return ciudadRepository.findById(id);
+    public boolean delete(Integer id){
+        boolean toret = false;
+        IdsDtos log = new IdsDtos(id, null, null);
+        Optional<Ciudad> byId = ciudadRepository.findById(id);
+        if(byId.isPresent()) {
+            ciudadRepository.deleteById(id);
+            logService.deleteLog(log, HttpStatus.OK.toString());
+            toret = true;
+        } else
+            logService.deleteLog(null, HttpStatus.NOT_FOUND.toString());
+        return toret;
     }
 }
