+++ demo/src/main/java/com/example/demo/services/SedeJJOOService.java
@@ -3,25 +3,27 @@ package com.example.demo.services;
 import com.example.demo.dtos.FullDtos.CiudadFullDtos;
 import com.example.demo.dtos.FullDtos.PaisFullDtos;
 import com.example.demo.dtos.FullDtos.SedeJJOOFullDtos;
+import com.example.demo.dtos.IdsDtos;
 import com.example.demo.dtos.SedeJJOODtos;
-import com.example.demo.entities.*;
+import com.example.demo.entities.Mysql.*;
 import com.example.demo.repository.Mysql.CiudadRepository;
 import com.example.demo.repository.Mysql.PaisRepository;
 import com.example.demo.repository.Mysql.SedeJJOORepository;
 import com.example.demo.repository.Mysql.TipoJJOORepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 @Transactional
 public class SedeJJOOService {
+
+    @Autowired
+    private LogService logService;
 
     @Autowired
     private SedeJJOORepository sedeJJOORepository;
@@ -35,7 +37,7 @@ public class SedeJJOOService {
     @Autowired
     private TipoJJOORepository tipoJJOORepository;
 
-    public Optional<List> findAll() {
+    public List findAll() {
         List<SedeJJOODtos> toret = new ArrayList<>();
         List<SedeJJOO> sedeJJOOList = sedeJJOORepository.findAll();
         Iterator<SedeJJOO> listIt = sedeJJOOList.iterator();
@@ -44,10 +46,14 @@ public class SedeJJOOService {
             SedeJJOODtos sedeJJOODtos = new SedeJJOODtos(sedeJJOO.getSedeJJOOPK().getAño(), sedeJJOO.getSedeJJOOPK().getTipoJJOO().getIdTipoJJOO(), sedeJJOO.getCiudadSede().getIdCiudad());
             toret.add(sedeJJOODtos);
         }
-        return Optional.of(toret);
+        if(toret.size() == 0)
+            logService.getLog(toret, HttpStatus.OK.toString());
+        else
+            logService.getLog(toret, HttpStatus.OK.toString());
+        return toret;
     }
 
-    public Optional<List> findAllFull() {
+    public List findAllFull() {
         List<SedeJJOOFullDtos> toret = new ArrayList<>();
         List<SedeJJOO> sedeJJOOList = sedeJJOORepository.findAll(new Sort(Sort.DEFAULT_DIRECTION, "sedeJJOOPK"));
         Iterator<SedeJJOO> listIt = sedeJJOOList.iterator();
@@ -58,7 +64,11 @@ public class SedeJJOOService {
             SedeJJOOFullDtos sedeJJOOFullDtos = new SedeJJOOFullDtos(sedeJJOO.getSedeJJOOPK().getAño(), sedeJJOO.getSedeJJOOPK().getTipoJJOO().getIdTipoJJOO(), ciudadFullDtos);
             toret.add(sedeJJOOFullDtos);
         }
-        return Optional.of(toret);
+        if(toret.size() == 0)
+            logService.getLog(toret, HttpStatus.OK.toString());
+        else
+            logService.getLog(toret, HttpStatus.OK.toString());
+        return toret;
     }
 
     public Optional save(SedeJJOODtos sedeJJOODtos) {
@@ -108,18 +118,32 @@ public class SedeJJOOService {
             if(sedeJJOOOptional.isPresent()) {
                 SedeJJOO sedeJJOO = (SedeJJOO) sedeJJOOOptional.get();
                 SedeJJOODtos sedeJJOODtos = new SedeJJOODtos(sedeJJOO.getSedeJJOOPK().getAño(), sedeJJOO.getSedeJJOOPK().getTipoJJOO().getIdTipoJJOO(), sedeJJOO.getCiudadSede().getIdCiudad());
+                List<SedeJJOODtos> log = new ArrayList<>();
+                log.add(sedeJJOODtos);
+                logService.getLog(log , HttpStatus.OK.toString());
                 return Optional.of(sedeJJOODtos);
             }
-            else
+            else {
+                logService.getLog(null, HttpStatus.NOT_FOUND.toString());
                 return sedeJJOOOptional;
-        }
-        else
+            }
+        } else {
+            logService.getLog(null, HttpStatus.NOT_FOUND.toString());
             return optionalTipoJJOO;
+        }
     }
 
-    public void deleteById(Integer id, Integer idTipoJJOO) {
-        TipoJJOO tipoJJOO = tipoJJOORepository.findById(idTipoJJOO).get();
-        SedeJJOOPK sedeJJOOPK = new SedeJJOOPK(id, tipoJJOO);
-        sedeJJOORepository.deleteById(sedeJJOOPK);
+    public boolean deleteById(Integer id, Integer idTipoJJOO) {
+        boolean toret = false;
+        if(findById(id, idTipoJJOO).isPresent()){
+            IdsDtos log = new IdsDtos(null, id, idTipoJJOO);
+            TipoJJOO tipoJJOO = tipoJJOORepository.findById(idTipoJJOO).get();
+            SedeJJOOPK sedeJJOOPK = new SedeJJOOPK(id, tipoJJOO);
+            sedeJJOORepository.deleteById(sedeJJOOPK);
+            logService.deleteLog(log, HttpStatus.OK.toString());
+            toret = true;
+        } else
+            logService.deleteLog(null, HttpStatus.NOT_FOUND.toString());
+        return toret;
     }
 }
